Other: Renamed the InlineEditor file to inlineeditor.js to match the naming convention. Closes #6.
