+++ tests/manual/inlineeditor.html
