+++ tests/manual/inlineeditor.md
